+++ packages/react-graphql/src/async/index.ts
@@ -1,0 +1,2 @@
+export {createAsyncQuery} from './query';
+export {createAsyncQueryComponent} from './component';